Modified admins.hbs
I believe that a lot of the information we want to show on the users page will not be able to be shown because these accounts are not going to be linked to individual people. The users panel will only exist to show the users, roles, and the ability to reset a person's password.